pyproject.toml
[project]
name = "kafka-python-getting-started"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kafka-python-ng>=2.2.3",
]

[tool.pyright]
reportMissingTypeStubs = false
